@@ -1,14 +1,13 @@
 ﻿<Project>
 
   <PropertyGroup>
-    <TargetFramework>net7.0-windows</TargetFramework>
+    <TargetFramework>net8.0-windows</TargetFramework>
   </PropertyGroup>
 
   <Import Sdk="Microsoft.NET.Sdk.WindowsDesktop" Project="Sdk.props" />
 
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFrameworks>$(TargetFrameworks);net7.0-windows</TargetFrameworks>
     <ApplicationManifest>app1.manifest</ApplicationManifest>
     <UseWPF>true</UseWPF>
     <RuntimeIdentifier>win-x64</RuntimeIdentifier>
@@ -17,12 +16,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Eto.Veldrid.Wpf\Eto.Veldrid.Wpf.csproj" />
+    <ProjectReference Include="..\Eto.Veldrid.Wpf\Eto.Veldrid.Wpf.csproj" />
     <ProjectReference Include="..\TestEtoVeldrid\TestEtoVeldrid.csproj" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Eto.Platform.Wpf" Version="2.8.2" />
+    <PackageReference Include="Eto.Platform.Wpf" Version="2.8.3" />
 
     <!-- Please note that version 3.6.0 is the newest version of this toolkit
     that has licensing terms permitting use in commercial projects.
