@@ -20,13 +20,13 @@ namespace Eto.Veldrid.Wpf
 			Control.Loaded += Control_Loaded;
 		}
 
-		public Swapchain CreateSwapchain()
+		public Swapchain? CreateSwapchain()
 		{
-			Swapchain swapchain;
+			Swapchain? swapchain;
 
 			if (Widget.Backend == GraphicsBackend.OpenGL)
 			{
-				swapchain = Widget.GraphicsDevice.MainSwapchain;
+				swapchain = Widget.GraphicsDevice?.MainSwapchain;
 			}
 			else
 			{
@@ -40,7 +40,7 @@ namespace Eto.Veldrid.Wpf
 					Marshal.GetHINSTANCE(typeof(VeldridSurface).Module));
 
 				var renderSize = RenderSize;
-				swapchain = Widget.GraphicsDevice.ResourceFactory.CreateSwapchain(
+				swapchain = Widget.GraphicsDevice?.ResourceFactory.CreateSwapchain(
 					new SwapchainDescription(
 						source,
 						(uint)renderSize.Width,
@@ -61,7 +61,7 @@ namespace Eto.Veldrid.Wpf
 			Widget.SizeChanged += Widget_SizeChanged;
 		}
 
-		private void Widget_SizeChanged(object sender, EventArgs e)
+		private void Widget_SizeChanged(object? sender, EventArgs e)
 		{
 			Callback.OnResize(Widget, new ResizeEventArgs(RenderSize));
 		}
